+++ src/controls/Toast/index.tsx
@@ -1,0 +1,17 @@
+
+// NOT A REAL COMPONENT
+// THIS FILE IS PURELY FOR DOCUMENTATION PURPOSES
+import React from 'react';
+
+/** 
+ * Toast is used to display messages in an overlay.
+ */
+class Toast extends React.Component {
+    static componentPathLine = `import { useToast } from '@ORIS/ui';`;
+
+    render() {
+        return null;
+    }
+}
+
+export default Toast;