+++ src/index.ts
@@ -30,6 +30,7 @@ export { default as SearchResult } from './controls/SearchResult';
 export { default as PersonSearchResult } from './controls/PersonSearchResult';
 export { default as TabList } from './controls/TabList';
 export { default as TabItem } from './controls/TabItem';
+export { useToast, ToastProvider } from './controls/Toast/hooks/useToast';
 export { default as Richtext } from './controls/Richtext';
 
 // Form