+++ CHANGELOG.md
@@ -9,8 +9,13 @@ Versions follow [Semantic Versioning](https://semver.org/) guidelines - given a 
 
 # 4.4.0 (In Progress)
 
-__Minor Backwards Breaking Changes:__
+__Deprecated Components:__
 
+* `Modal` (superseded by PrimeReact `Dialog`)
+
+__Breaking Changes:__
+
+* Potential breaking changes in all PrimeReact components due to upgrade to version 6.3.0 (see [\#1738](https://github.com/primefaces/primereact/issues/1738), [\#1566](https://github.com/primefaces/primereact/issues/1566), and [\#1877](https://github.com/primefaces/primereact/issues/1877))
 * `useSearchProvider` - Rename `results` to `response` and `setResults` to `setResponse` in the returned object
 
 __New Features:__
@@ -20,8 +25,10 @@ __New Features:__
 * `SyncSearchWithURL` updated to include pagination support.
 * Added `Paginator` component to be used within a `SearchProvider`.
 * Added support for parameters in `JsonApi` Search driver.
-* Added `Avatar` component with https://opic.osu.edu integration
-* Added `EmailLink` component
+* Added `Avatar` component with https://opic.osu.edu integration.
+* Added `EmailLink` component.
+* PrimeReact updated from version 5.0.0 to version 6.3.0.
+    * New PrimeReact component support in ORIS/ui: `Dialog` (replaces Modal), `Toast` (via `useToast` hook and `ToastProvider`), and `Skeleton`.
 
 __Bug Fixes:__